extractor-typin-pdf-microservice/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
minio==7.2.5
pymupdf==1.23.26
sentence-transformers==2.5.1
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
pydantic==2.6.3
python-multipart==0.0.9
psycopg2-binary==2.9.9
qdrant-client==1.8.0
